fireo/requirements.txt
filelock          3.12.2
Jinja2            3.1.2
MarkupSafe        2.1.3
mpmath            1.3.0
networkx          3.1
pip               22.0.4
PyYAML            6.0
setuptools        58.1.0
sympy             1.12
torch             2.0.1
typing_extensions 4.7.1